Cleanup - add svg_to_png export to label_builder.js
