+++ routes/labels.js
@@ -3,16 +3,6 @@ const router = express.Router();
 const fs = require('fs')
 const path = require('path');
 const svg_builder = require('../utils/label_builder.js');
-const { Resvg } = require('@resvg/resvg-js');
-
-const svg_opts = {
-    font: {
-        fontFiles: [path.join(__dirname, '../utils/fonts/Arimo-Regular.ttf'), 
-            path.join(__dirname, '../utils/fonts/Arimo-Bold.ttf'), 
-        ], // font files to use
-        defaultFontFamily: 'Arimo' // font to use for SVG to PNG conversion
-    }
-}
 
 const render_svg = async (req, res, svg_maker, callback) => {
     try {
@@ -75,7 +65,7 @@ router.get('/:participant_id/:id/png', function (req, res) {
                     res.status(500).send(err);
                     return;
                 }
-                const png_buffer = new Resvg(svg,svg_opts).render().asPng();
+                const png_buffer = svg_builder.svg_to_png(svg);
                 res.setHeader('Content-disposition', 'attachment; filename=Energy Rating Label - ' + pump.rating_id + '.png');
                 res.setHeader('Content-Type', 'image/png');
                 res.setHeader('Content-Length', png_buffer.length);
@@ -91,7 +81,7 @@ router.get('/:participant_id/:id/png-sm', function (req, res) {
                     res.status(500).send(err);
                     return;
                 }
-                const png_buffer = new Resvg(svg,svg_opts).render().asPng();
+                const png_buffer = svg_builder.svg_to_png(svg);
                 res.setHeader('Content-disposition', 'attachment; filename=Energy Rating Label (sm) - ' + pump.rating_id + '.png');
                 res.setHeader('Content-Type', 'image/png');
                 res.setHeader('Content-Length', png_buffer.length);
@@ -119,7 +109,7 @@ router.get('/:participant_id/:id/qr/png', function (req, res) {
                     res.status(500).send(err);
                     return;
                 }
-                const png_buffer = new Resvg(svg,svg_opts).render().asPng();
+                const png_buffer = svg_builder.svg_to_png(svg);
                 res.setHeader('Content-disposition', 'attachment; filename=Energy Rating QR - ' + pump.rating_id + '.png');
                 res.setHeader('Content-Type', 'image/png');
                 res.setHeader('Content-Length', png_buffer.length);
